Moved flags logic out of publication component into their own
