+++ api/src/components/flag/schema/index.ts
@@ -1,0 +1,2 @@
+export { default as createFlag } from './createFlag';
+export { default as createFlagComment } from './createFlagComment';
