+++ api/src/components/publication/routes.ts
@@ -37,20 +37,3 @@ export const updateStatus = middy(publicationController.updateStatus)
     .use(middleware.httpJsonBodyParser())
     .use(middleware.authentication())
     .use(middleware.validator(publicationSchema.updateStatus, 'pathParameters'));
-
-export const createFlag = middy(publicationController.createFlag)
-    .use(middleware.doNotWaitForEmptyEventLoop({ runOnError: true, runOnBefore: true, runOnAfter: true }))
-    .use(middleware.httpJsonBodyParser())
-    .use(middleware.authentication())
-    .use(middleware.validator(publicationSchema.createFlag, 'body'));
-
-export const createFlagComment = middy(publicationController.createFlagComment)
-    .use(middleware.doNotWaitForEmptyEventLoop({ runOnError: true, runOnBefore: true, runOnAfter: true }))
-    .use(middleware.httpJsonBodyParser())
-    .use(middleware.authentication())
-    .use(middleware.validator(publicationSchema.createFlagComment, 'body'));
-
-export const resolveFlag = middy(publicationController.resolveFlag)
-    .use(middleware.doNotWaitForEmptyEventLoop({ runOnError: true, runOnBefore: true, runOnAfter: true }))
-    .use(middleware.httpJsonBodyParser())
-    .use(middleware.authentication());
