+++ api/src/components/publication/schema/index.ts
@@ -2,5 +2,3 @@ export { default as create } from './create';
 export { default as updateStatus } from './updateStatus';
 export { default as getAll } from './getAll';
 export { default as update } from './update';
-export { default as createFlag } from './createFlag';
-export { default as createFlagComment } from './createFlagComment';
