+++ api/tsconfig.json
@@ -54,6 +54,9 @@
             "coauthor/*": [
                 "src/components/coauthor/*"
             ],
+            "flag/*": [
+                "src/components/flag/*"
+            ],
             "lib/*": [
                 "src/lib/*"
             ],
